test: update test for better code coverage
